+++ src/views/login.ts
@@ -8,31 +8,34 @@ import { AuthenticatedStatus, LoginInfo } from '../interfaces';
   template: `<h2>Login</h2>
   <form #form="ngForm" (ngSubmit)="onSubmit(form.value)">
     <p class="error" *ngIf="error">{{ error }}</p>
-    <p><label>Login <input type="text" name="login" ngModel /></label></p>
-    <p><label>Password <input type="password" name="password" ngModel /></label></p>
-    <input type="submit" value="Login" />
+    <p><label>Login <input type="text" name="login" ngModel/></label></p>
+    <p><label>Password <input type="password" name="password" ngModel/></label></p>
+    <input type="submit" value="Login"/>
   </form>`
 })
 export class LoginView implements OnInit {
   error: string = '';
+  isLogged: boolean;
 
-  constructor(
-    public services: Services,
-  ) { }
+  constructor(public services: Services,) {
+  }
 
   ngOnInit() {
     this.services.authentication.isAuthenticated
       .subscribe((auth: AuthenticatedStatus) => {
-        if (auth.state) {
-          this.services.traverser.traverse(
-            this.services.traverser.target.getValue().contextPath);
-        }
-        this.error = auth.error || '';
+        this.isLogged = auth.state;
       });
   }
 
   onSubmit(data: LoginInfo) {
-    this.services.authentication.login(data.login, data.password);
+    this.services.authentication.login(data.login, data.password)
+      .subscribe(() => {
+        this.services.traverser.traverse(
+          this.services.traverser.target.getValue().contextPath);
+        this.error = '';
+      }, (error: Error) => {
+        this.error = error.message;
+      });
   }
 
 }